LayoutTests/imported/w3c/web-platform-tests/html/browsers/history/the-location-interface/document_location-expected.txt


PASS document not in a browsing context
PASS document.location equals window.location
FAIL Attribute getter deduplication assert_equals: expected function "function location() {
    [native code]
}" but got function "function location() {
    [native code]
}"
FAIL Attribute setter deduplication assert_equals: expected function "function location() {
    [native code]
}" but got function "function location() {
    [native code]
}"
